+++ Assets/_Workspace/Scripts/Interface/ITakeDamage.cs
@@ -1,5 +1,5 @@
 public interface ITakeDamage
 {
-    void TakeDamage(int value, bool headShoot);
+    void TakeDamage(float value, bool headShoot);
     void Dead(bool headShoot); 
 }
